requirements.txt
alembic==1.10.3
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Cors==3.0.10
flask-form==1.0.0
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
markup==0.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
Safe==0.4
six==1.15.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
